Change variable name from "JPRIME" to "isPrime"
Make it more understandable to the reader by changing the name of the
variable, and changing the way it is used in the method. For example,
when a number is prime, it should go in the while loop since the number
IS prime, however the way it was coded before made it hard to
understand since it was a double negative. (Before, the line read
itself as while the number isnt (isntPrime)) This way, it is much
clearer what the coder meant